Interim commit - climbing motor control
Code for controling the two climbing motors (Motor5 & Motor6) have been
added. A new PID controller object (pid_controller_1) has been
instantiated and is being used to look at the counter feedback from
Motor5 and Motor6 and is fed back in the form of an adjustment to the
speed of one of the Motors.