Autonomous mode changes
added code to reset the direction of the jaguars. the settings were
carried over from teleoperated mode and had to be reset. We also
modified the speed and curve parameters as the robot behaves differently
on carpet. There is also a for loop added to ease the robot up to full
speed.